Solved 2270 Number of Ways to Split Array with Optimize

+++ 01 badge 2025/03_2270. Number of Ways to Split Array/solution1-bruteforce.cpp
@@ -5,7 +5,7 @@ using namespace std;
 class Solution {
 public:
     int waysToSplitArray(vector<int>& nums) {
-        long left=0, right=0;
+        long long left=0, right=0;
         int res=0;
         for(int i=0; i<nums.size(); i++) right += nums[i];
         for(int i=0; i<nums.size()-1; i++){
